content/base/test/test_bug1025933.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1025933
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1025933</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1025933 **/

  SimpleTest.waitForExplicitFinish();

  function test() {
    var s = document.getElementById("host").createShadowRoot();
    s.innerHTML = '<div style="width:100px;height:100px;background:red"></div>';
    var el = s.firstElementChild;
    is(el.clientWidth, 100);
    is(el.clientHeight, 100);
    SimpleTest.finish();
  }

  </script>
</head>
<body onload="test()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1025933">Mozilla Bug 1025933</a>
<p id="display"></p>
<div id="content">
  <div id="host"></div>
</div>
<pre id="test">
</pre>
</body>
</html>
